consolidating the use of hashes
