+++ datajoint/hash.py
@@ -1,6 +1,15 @@
 import hashlib
 import base64
 
+def key_hash(key):
+    """
+    32-byte hash used for lookup of primary keys of jobs
+    """
+    hashed = hashlib.md5()
+    for k, v in sorted(key.items()):
+        hashed.update(str(v).encode())
+    return hashed.hexdigest()
+
 
 def to_ascii(byte_string):
     """
@@ -10,25 +19,24 @@ def to_ascii(byte_string):
     return base64.b64encode(byte_string, b'-_').decode()
 
 
-def long_hash(buffer):
+def long_hash(*buffers):
     """
     :param buffer: a binary buffer (e.g. serialized blob)
     :return: 43-character base64 ASCII rendition SHA-256
     """
-    return to_ascii(hashlib.sha256(buffer).digest())[0:43]
+    hashed = hashlib.sha256()
+    for buffer in buffers:
+        hashed.update(buffer)
+    return to_ascii(hashed.digest())[0:43]
 
 
-def short_hash(buffer):
+def short_hash(*buffers):
     """
     :param buffer: a binary buffer (e.g. serialized blob)
     :return: the first 8 characters of base64 ASCII rendition SHA-1
     """
-    return to_ascii(hashlib.sha1(buffer).digest())[:8]
+    hashed = hashlib.sha1()
+    for buffer in buffers:
+        hashed.update(buffer)
+    return to_ascii(hashed.digest())[:8]
 
-
-# def filehash(filename):
-#     s = hashlib.sha256()
-#     with open(filename, 'rb') as f:
-#         for block in iter(lambda: f.read(65536), b''):
-#             s.update(block)
-#     return to_ascii(s.digest())
